Add modularity
